refactor: separate `util` library target
# Motivation

Improves separation of concerns between packages, which should improve
readability.
