+++ wkbre2/gameset/cameraPath.cpp
@@ -4,7 +4,7 @@
 
 #include "cameraPath.h"
 #include "gameset.h"
-#include "../util/GSFileParser.h"
+#include "util/GSFileParser.h"
 
 void CameraPath::parse(GSFileParser& gsf, GameSet& gs)
 {
