+++ wkbre2/gameset/gameset.h
@@ -8,16 +8,16 @@
 #include <set>
 #include <memory>
 #include <string>
-#include "../util/IndexedStringList.h"
+#include "util/IndexedStringList.h"
 #include "../tags.h"
 #include "GameObjBlueprint.h"
 #include "values.h"
 #include "actions.h"
-#include "../util/DynArray.h"
+#include "util/DynArray.h"
 #include "command.h"
 #include "../Model.h"
 #include "OrderBlueprint.h"
-#include "../util/StriCompare.h"
+#include "util/StriCompare.h"
 #include "reaction.h"
 #include "ObjectCreation.h"
 #include "condition.h"
