+++ wkbre2/gameset/reaction.cpp
@@ -3,7 +3,7 @@
 // Licensed under the GNU General Public License 3
 
 #include "reaction.h"
-#include "../util/GSFileParser.h"
+#include "util/GSFileParser.h"
 #include "gameset.h"
 #include "actions.h"
 #include "../server.h"
