+++ wkbre2/gfx/renderer.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include <cstdint>
-#include "../util/vecmat.h"
+#include "util/vecmat.h"
 
 struct Bitmap;
