+++ wkbre2/util/include/util/IndexedStringList.h
