+++ wkbre2/util/include/util/StriCompare.h
