+++ wkbre2/util/include/util/vecmat.h
