+++ wkbre2/util/util.cpp
@@ -2,7 +2,7 @@
 // (C) 2021 AdrienTD
 // Licensed under the GNU General Public License 3
 
-#include "util.h"
+#include "util/util.h"
 #include <cstdlib>
 #include <cstdarg>
 #include <cstdio>
